+++ hw2-coworking-service/docker-compose.yml
@@ -1,0 +1,15 @@
+version: "3.9"
+services:
+  postgres:
+    image: postgres:latest
+    environment:
+      POSTGRES_DB: "coworkingdb"
+      POSTGRES_USER: "coworkinguser"
+      POSTGRES_PASSWORD: "coworkingpassword"
+      PGDATA: "/var/lib/postgresql/data/pgdata"
+    volumes:
+      - C:\Users\After\Documents\JavaProject\y_lab-homework\hw2-coworking-service\src\main\resources\db.changelog\init\init-schema.sql:/docker-entrypoint-initdb.d
+    ports:
+      - "5432:5432"
+
+
