+++ hw2-coworking-service/src/main/java/com/coworkingservice/CoworkingService.java
@@ -1,12 +1,7 @@
 package com.coworkingservice;
 
-
 import com.coworkingservice.entity.*;
-import com.coworkingservice.fabric.EntityFabric;
-import com.coworkingservice.memorydb.*;
 import com.coworkingservice.service.ScannerSingleton;
-import com.coworkingservice.service.booking.BookingSlot;
-import com.coworkingservice.service.verify.VerifyDate;
 
 import java.util.Map;
 
@@ -18,16 +13,16 @@ public class CoworkingService {
         ScannerSingleton.getInstance().getScanner().close();
     }
 
-    private static void generateStartingData(){
+    private static void generateStartingData() {
         long roomId = 1;
         Map<Long, Room> testMap = MemoryDB.getInstance().getRoomMapTable();
         Map<Credential, Person> testMap2 = MemoryDB.getInstance().getPersonMapTable();
-        for(int i = 0; i < 10; i++){
-            testMap.put(roomId,new WorkplaceRoom(roomId++));
+        for (int i = 0; i < 10; i++) {
+            testMap.put(roomId, new WorkplaceRoom(roomId++));
         }
-        for(int i = 0; i < 5; i++){
-            testMap.put(roomId,new ConferenceRoom(roomId++));
+        for (int i = 0; i < 5; i++) {
+            testMap.put(roomId, new ConferenceRoom(roomId++));
         }
-        testMap2.put(new Credential("login","password"), new Tenant("Raw","User"));
+        testMap2.put(new Credential("login", "password"), new Tenant("Raw", "User"));
     }
 }
