+++ hw2-coworking-service/src/main/java/com/coworkingservice/memorydb/PersonCRUD.java
@@ -1,9 +1,9 @@
 package com.coworkingservice.memorydb;
 
+import com.coworkingservice.MemoryDB;
 import com.coworkingservice.entity.Credential;
 import com.coworkingservice.entity.Person;
 
-import com.coworkingservice.fabric.EntityFabric;
 import com.coworkingservice.fabric.EntityFamilyFabric;
 import com.coworkingservice.service.verify.Registration;
 import com.coworkingservice.service.verify.PersonRegistration;
