+++ hw2-coworking-service/src/main/java/com/coworkingservice/memorydb/ReservedSlotsCRUD.java
@@ -1,6 +1,7 @@
 package com.coworkingservice.memorydb;
 
 
+import com.coworkingservice.MemoryDB;
 import com.coworkingservice.entity.Person;
 import com.coworkingservice.entity.Room;
 import com.coworkingservice.entity.Slot;
