+++ hw2-coworking-service/src/main/java/com/coworkingservice/memorydb/RoomCRUD.java
@@ -1,6 +1,7 @@
 package com.coworkingservice.memorydb;
 
 
+import com.coworkingservice.MemoryDB;
 import com.coworkingservice.entity.Room;
