+++ hw2-coworking-service/src/main/java/com/coworkingservice/service/booking/BookingSlot.java
@@ -1,7 +1,7 @@
 package com.coworkingservice.service.booking;
 
 import com.coworkingservice.entity.Slot;
-import com.coworkingservice.memorydb.MemoryDB;
+import com.coworkingservice.MemoryDB;
 import com.coworkingservice.service.verify.VerifyDate;
 
 import java.util.List;
