+++ hw2-coworking-service/src/main/resources/db.changelog/changelog.xml
@@ -1,6 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <databaseChangeLog xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
                    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                   xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.6.xsd">
-
+                   xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-4.5.xsd">
+    <include file="create-schema-entity.xml" relativeToChangelogFile="true"/>
+    <include file="create-table-person.xml" relativeToChangelogFile="true"/>
+    <include file="create-table-room.xml" relativeToChangelogFile="true"/>
+    <include file="create-table-slot.xml" relativeToChangelogFile="true"/>
+    <include file="insert-room-table.xml" relativeToChangelogFile="true"/>
+    <include file="insert-person-table.xml" relativeToChangelogFile="true"/>
 </databaseChangeLog>
