+++ hw2-coworking-service/src/main/resources/db.changelog/create-schema-entity.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<databaseChangeLog xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
+                   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                   xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-4.5.xsd">
+    <changeSet id="create-schema-entity" author="Georgiy">
+        <sql>
+            CREATE SCHEMA IF NOT EXISTS entity;
+        </sql>
+    </changeSet>
+</databaseChangeLog>
