+++ hw2-coworking-service/src/main/resources/db.changelog/create-table-person.xml
@@ -1,16 +1,16 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <databaseChangeLog xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
                    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                   xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.6.xsd">
+                   xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-4.5.xsd">
     <changeSet id="create-table-person" author="Georgiy">
-        <createTable tableName="person">
+        <createTable tableName="person" schemaName="entity">
             <column autoIncrement="true" name="id" type="SERIAL">
                 <constraints primaryKey="true" nullable="false" primaryKeyName="person_pkey"/>
             </column>
-            <column name="firstname">
+            <column name="firstname" type="varchar(255)">
                 <constraints nullable="false"/>
             </column>
-            <column name="lastname">
+            <column name="lastname" type="varchar(255)">
                 <constraints nullable="false"/>
             </column>
         </createTable>
