+++ hw2-coworking-service/src/main/resources/db.changelog/init/init-schema.sql
@@ -1,0 +1,1 @@
+CREATE SCHEMA IF NOT EXISTS service;
