+++ hw2-coworking-service/src/main/resources/db.changelog/liquibase.properties
@@ -1,4 +1,5 @@
-changeLogFile=changelog.xml
-url=jdbc:postgresql://localhost:5432/postgres
-username=postgres
-password=admin
+changeLogFile=src/main/resources/db.changelog/changelog.xml
+url=jdbc:postgresql://localhost:5432/coworkingdb
+username=coworkinguser
+password=coworkingpassword
+liquibaseSchemaName=service
