+++ hw2-coworking-service/src/test/java/com/coworkingservice/RoomCrudTest.java
@@ -4,10 +4,8 @@ import com.coworkingservice.entity.ConferenceRoom;
 import com.coworkingservice.entity.Room;
 import com.coworkingservice.entity.WorkplaceRoom;
 import com.coworkingservice.fabric.EntityFamilyFabric;
-import com.coworkingservice.memorydb.MemoryDB;
 import com.coworkingservice.memorydb.RoomCRUD;
 import org.junit.After;
-import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.jupiter.api.Assertions;
